put isLowerCase & isUpperCase into regexp files
