+++ src/is/index.js
@@ -1,7 +1,9 @@
 import types from './types';
 import arithmetic from './arithmetic';
+import regexp from './regexp';
 
 export default {
     ...types,
     ...arithmetic,
+    ...regexp,
 };
