+++ test/is/types.test.js
@@ -197,27 +197,3 @@ describe('isWindowObject', () => {
     });
     testFalsyWithNullable(is.windowObject);
 });
-
-describe('isLowerCase', () => {
-    test('returns true', () => {
-        expect(is.lowerCase('abc')).toBeTruthy();
-        expect(is.lowerCase('abc 123')).toBeTruthy();
-    });
-    test('returns false', () => {
-        expect(is.lowerCase(1)).toBeFalsy();
-        expect(is.lowerCase('Abc')).toBeFalsy();
-        expect(is.lowerCase('ABC')).toBeFalsy();
-    });
-});
-
-describe('isUpperCase', () => {
-    test('returns true', () => {
-        expect(is.upperCase('ABC')).toBeTruthy();
-        expect(is.upperCase('ABC 123')).toBeTruthy();
-    });
-    test('returns false', () => {
-        expect(is.upperCase(1)).toBeFalsy();
-        expect(is.upperCase('aBC')).toBeFalsy();
-        expect(is.upperCase('abc')).toBeFalsy();
-    });
-});
